+++ CHANGELOG.md
@@ -10,6 +10,7 @@ All notable changes to this project will be documented in this file.
 - The `mongodb` processor and output default `write_concern.w_timeout` empty value no longer causes configuration issues.
 - Field `message_name` added to the logger config.
 - The `amqp_1` input and output should no longer spam logs with timeout errors during graceful termination.
+- Fixed a potential crash when the `contains` bloblang method was used to compare complex types.
 
 ### Changed